requirements.txt
python-twitter
requests
unidecode
telebot
telethon